requirements.txt
# Core dependencies
pydantic>=2.0.0
typing-extensions>=4.5.0

# LangChain and LangGraph (required for agent workflows)
langchain>=0.3.0
langchain-core>=0.3.0
langgraph>=0.2.0
langchain-openai>=0.2.0
langchain-community>=0.3.0

# OpenAI (optional - if using OpenAI models)
openai>=1.0.0

# Anthropic (optional - if using Claude models)
anthropic>=0.34.0

# Document processing
python-dotenv>=1.0.0

# Additional utilities
requests>=2.31.0
